requirements.txt
amqp==5.3.1
asgiref==3.8.1
bcrypt==4.3.0
billiard==4.2.1
Brotli==1.1.0
celery==5.5.2
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==45.0.3
cssselect2==0.8.0
Django==5.1.7
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-filter==23.5
django-restframework==0.0.1
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
drf-spectacular==0.27.2
et_xmlfile==2.0.0
fonttools==4.58.4
gunicorn==23.0.0
inflection==0.5.1
kombu==5.5.3
mysqlclient==2.2.7
numpy==2.3.0
openpyxl==3.1.5
packaging==24.2
pandas==2.3.0
paramiko==3.5.1
pillow==11.2.1
prompt_toolkit==3.0.51
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pymongo==4.13.2
PyNaCl==1.5.0
pyphen==0.17.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
reportlab==4.4.2
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1